% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6018.R
\docType{package}
\name{R6018}
\alias{R6018}
\title{R6018: A package for SYS-6018}
\description{
The R6018 package is designed to help students in SYS-6018. It provides some
helpful functions, formats, and datasets.
}
\section{R6018 functions}{

\code{\link{digits}}
}

\section{R6018 Formats}{

\code{\link{homework}}
}

\section{R6018 Datasets}{

TODO
}

